Renames ops.rs to crud_ops.rs and adds safe guard to write_to_stream
